[#2] Refactor: 비즈니스 로직 hooks 로 리펙터리
